pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-validator"
version = "0.1.0"
description = "Agent validator Service for Matrix EcoSystem"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }

dependencies = [
    "fastapi==0.111.0",
    "groq>=0.32.0",
    "uvicorn[standard]==0.29.0",
    "httpx==0.28.1",
    "pydantic>=2.7.1",
    "python-json-logger==2.0.7",
    "cachetools==5.3.3",
    "huggingface-hub==0.23.0",
    "sentence-transformers==2.7.0",
    "faiss-cpu==1.8.0",
    "numpy==1.26.4",
    "orjson==3.10.3",
    "pyyaml==6.0.1",
    "tenacity==8.2.3",
    "python-dotenv==1.0.1",
    "google-genai>=1.39.1",

    # --- Added for a2a-validator ---
    "a2a-sdk[http-server]>=0.3.0",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.2.1",
    "typing-extensions>=4.12.2",
    # FIX: Ensure all standard websocket dependencies are included
    "python-socketio[asyncio_standard]>=5.11.0",
    "jinja2>=3.1.2",
    "bleach>=6.2.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]